[WIP] A bit more of reorganization
Moved `MatchProperties` from `prospective::match_properties` to `prospective`.
