+++ src/core/prospective/context.rs
